refactor: extract overview QuestionsSection component
